requirements.txt
pytest==8.3.3
python-dotenv==1.0.1
seleniumbase==4.33.1

